feat(BUX-612): cleaning the index.ts file
